Committed fixes for bugs found while writing the initial "failed due to rsl problems" test cases for GRAM - fork job manager script (fork.in)   error on non-literal directory attribute   use correct error (BAD_DIRECTORY) for directory which doesn't exist   error on invalid enironment RSL (non-pairs)   error on non-integer count RSL   error JOBTYPE_NOT_SUPPORTED for bogus job types   error on invalid executable (non-literal)   error on no-such-executable   error on non-executable executable   error on invalid stdin (non-literal)   correctly deal with arguments when they are undefined - job manager.h   add prototypes for new functions - globus_gram_job_manager_output.c   return real errors (not GLOBUS_FAILURE) when failed to open stdout or       stderr file - globus_gram_job_manager_request.c   don't lose logfile name when $(HOME) is not writable (bugz #125) - globus_gram_job_manager_rsl.c   add new function (globus_gram_job_manager_rsl_need_restart)   error on NULL RSL   erorr on uncanonicalizable RSL   error propagation from creation of output destinations   error on invalid_count   error on invalid myjob   error on invalid dryrun   error on invalid save_state   error on invalid two_phase_commit   error on invalid remote_io_url   fix leak on evalation_value and return errors when eval fails. - globus_gram_job_manager_script.c   set state to FAILED instead of DONE when a dryrun occurs - globus_gram_job_manager_staging.c   set error that matches staging file type (STDIN,EXE,FILE_STAGE_IN)   fix evaluating staging URLS   detect difference from evaluation failure and invalid RSL for staging   in, out, and err values. - globus_gram_job_manager_state.c   error on invalid restart parameter (non-literal)   error on invalid cache parameter (non-literal)   error on invalid cacahe location value   catch dry-run error when doing a submit attempt - globus_gram_job_manager_validate.c   error on non-GRAM RSL   allow case-insensitive and "_"-insensitive comparisons for attributes       in validation file   return proper error for attributes which can't be used in a validation       context   return RSL attribute-specific errors when non-literal RSL values should       match an enumerated value (with a generic error for scheduler-specific       RSL attributes)   return RSL attribute-specific errors when RSL doesn't match RSL values       (with a generic error for scheduler-specific RSL attributes) - JobDescription.pm   handle array and scalar contexts differently; makes it possible to       detect missing vs invalid RSL values (and the old scalar() calls were       wrong) - JobManager.pm   error on bad format for file_cleanup RSL values

+++ gram/jobmanager/source/scripts/JobManager.pm
@@ -367,6 +367,10 @@ sub file_cleanup
 	{
 	    next;
 	}
+	if(ref($_))
+	{
+	    return Globus::GRAM::Error::RSL_FILE_CLEANUP();
+	}
 	$self->log("Removing $_");
 
 	unlink($_);
